+++ GloryECS/GloryReflectStatic/FieldData.h
@@ -1,9 +1,10 @@
 #pragma once
+#include <stdint.h>
 
 namespace GloryReflect
 {
-	using FieldType = size_t;
-	using ElementType = size_t;
+	using FieldType = uint32_t;
+	using ElementType = uint32_t;
 
 	struct FieldData
 	{
